requirements.txt
langchain
langchain-openai
chromadb
pypdf
python-dotenv
fastapi
uvicorn
